Lesson 13

+++ APLesson_13.2/Ticket.java
@@ -14,7 +14,7 @@ public abstract class Ticket
 		return (int)(Math.random() * 10000000) + 1; 
 	}
 	
-	public abstract int getPrice();
+	public abstract double getPrice();
 	
 	public String toString()
 	{
